Add files via upload
Guess the Number is a simple but fun guessing game, in which you must guess a randomly chosen number by the computer. You have several tries to find the right number, and each round, the computer will tell you if your guess is too high or too low.  The game is developed in python, and is easy to play for users of all ages. Try it out now and test your guessing skills!